==> initial_etl/3_data_preparation_db.ipynb <==
import pandas as pd
import numpy as np

# Accidents

accidents_df = pd.read_json("shapefiles/accidents/accidents_2015-2022.json")

# Extract date info
    # The unix time is UTC time, which is five hours ahead of Bogotá
    # https://thispointer.com/subtract-hours-from-datetime-in-python/
accidents_df["FECHA_HORA_ACC_r"] = pd.to_datetime(accidents_df["FECHA_HORA_ACC"], unit = "ms") - pd.DateOffset(hours = 5)

    # The following approach is more formal, but leaves the column with a final "flag" indicating the time difference w.r.t. UTC
#accidents_df["FECHA_HORA_ACC_r"] = pd.to_datetime(accidents_df["FECHA_HORA_ACC"], unit = "ms").dt.tz_localize("UTC").\
#    dt.tz_convert("America/Bogota")

accidents_df["ANO_OCURRENCIA_ACC"] = accidents_df["FECHA_HORA_ACC_r"].dt.year

    # Install the Spanish locale before running the following code
    # https://serverpilot.io/docs/how-to-install-locales/
accidents_df["MES_OCURRENCIA_ACC"] = accidents_df["FECHA_HORA_ACC_r"].dt.month_name(locale = "es_ES.UTF-8")
accidents_df["MES_NRO_OCURRENCIA_ACC"] = accidents_df["FECHA_HORA_ACC_r"].dt.month
accidents_df["DIA_OCURRENCIA_ACC"] = accidents_df["FECHA_HORA_ACC_r"].dt.day_name(locale = "es_ES.UTF-8")

    # 1 is Monday, 2 is Tuesday and so on
accidents_df["DIA_NRO_OCURRENCIA_ACC"] = accidents_df["FECHA_HORA_ACC_r"].dt.dayofweek + 1

accidents_df["DIA_MES_OCURRENCIA_ACC"] = accidents_df["FECHA_HORA_ACC_r"].dt.day
accidents_df["HORA_OCURRENCIA_ACC"] = accidents_df["FECHA_HORA_ACC_r"].dt.hour

# Uppercase month and day of week to keep consistency with original format
accidents_df["MES_OCURRENCIA_ACC"] = accidents_df["MES_OCURRENCIA_ACC"].str.upper()
accidents_df["DIA_OCURRENCIA_ACC"] = accidents_df["DIA_OCURRENCIA_ACC"].str.upper()

# Rename the longitude and latitude columns
accidents_df.rename(columns = {"geometry.x": "LONGITUDE", "geometry.y": "LATITUDE"}, inplace = True)

# CIV and PK_CALZADA NaN treatment: fill with 0s and then change their type to int
accidents_df["CIV"] = accidents_df["CIV"].fillna(0)
accidents_df["PK_CALZADA"] = accidents_df["PK_CALZADA"].fillna(0)
accidents_df["CIV"] = accidents_df["CIV"].astype(int)
accidents_df["PK_CALZADA"] = accidents_df["PK_CALZADA"].astype(int)

accidents_df.to_json("shapefiles/accidents_2015-2022_r1.json")

accidents_df.info()

accidents_df.head()

# Injured people

injured_people_df = pd.read_json("shapefiles/injured/injured_people_2015-2022.json")

# Extract date info
    # The unix time is UTC time, which is five hours ahead of Bogotá
    # https://thispointer.com/subtract-hours-from-datetime-in-python/
injured_people_df["FECHA_HORA_ACC_r"] = pd.to_datetime(injured_people_df["FECHA_HORA_ACC"], unit = "ms") - pd.DateOffset(hours = 5)

    # The following approach is more formal, but leaves the column with a final "flag" indicating the time difference w.r.t. UTC
#injured_people_df["FECHA_HORA_ACC_r"] = pd.to_datetime(injured_people_df["FECHA_HORA_ACC"], unit = "ms").dt.tz_localize("UTC").\
#    dt.tz_convert("America/Bogota")

injured_people_df["ANO_OCURRENCIA_ACC"] = injured_people_df["FECHA_HORA_ACC_r"].dt.year

    # Install the Spanish locale before running the following code
    # https://serverpilot.io/docs/how-to-install-locales/
injured_people_df["MES_OCURRENCIA_ACC"] = injured_people_df["FECHA_HORA_ACC_r"].dt.month_name(locale = "es_ES.UTF-8")
injured_people_df["MES_NRO_OCURRENCIA_ACC"] = injured_people_df["FECHA_HORA_ACC_r"].dt.month
injured_people_df["DIA_OCURRENCIA_ACC"] = injured_people_df["FECHA_HORA_ACC_r"].dt.day_name(locale = "es_ES.UTF-8")

    # 1 is Monday, 2 is Tuesday and so on
injured_people_df["DIA_NRO_OCURRENCIA_ACC"] = injured_people_df["FECHA_HORA_ACC_r"].dt.dayofweek + 1

injured_people_df["DIA_MES_OCURRENCIA_ACC"] = injured_people_df["FECHA_HORA_ACC_r"].dt.day
injured_people_df["HORA_OCURRENCIA_ACC"] = injured_people_df["FECHA_HORA_ACC_r"].dt.hour

# Uppercase month and day of week to keep consistency with original format
injured_people_df["MES_OCURRENCIA_ACC"] = injured_people_df["MES_OCURRENCIA_ACC"].str.upper()
injured_people_df["DIA_OCURRENCIA_ACC"] = injured_people_df["DIA_OCURRENCIA_ACC"].str.upper()

# Rename the longitude and latitude columns
injured_people_df.rename(columns = {"geometry.x": "LONGITUDE", "geometry.y": "LATITUDE"}, inplace = True)

# GENERO NaN treatment: fill with "SIN INFORMACION"
injured_people_df["GENERO"] = injured_people_df["GENERO"].fillna("SIN INFORMACION")

# EDAD NaN treatment: since we don't have information on what a zero represents (is it a baby that hasn't turned one year yet or
# a null value?), we leave this feature as it is

injured_people_df.to_json("shapefiles/injured_people_2015-2022_r1.json")

injured_people_df.info()

injured_people_df.head()

# Killed people

killed_people_df = pd.read_json("shapefiles/killed/killed_people_2015-2022_r1.json")

# Extract date info
    # The unix time is UTC time, which is five hours ahead of Bogotá
    # https://thispointer.com/subtract-hours-from-datetime-in-python/
killed_people_df["FECHA_HORA_ACC_r"] = pd.to_datetime(killed_people_df["FECHA_HORA_ACC"], unit = "ms") - pd.DateOffset(hours = 5)

    # The following approach is more formal, but leaves the column with a final "flag" indicating the time difference w.r.t. UTC
#killed_people_df["FECHA_HORA_ACC_r"] = pd.to_datetime(killed_people_df["FECHA_HORA_ACC"], unit = "ms").dt.tz_localize("UTC").\
#    dt.tz_convert("America/Bogota")

killed_people_df["ANO_OCURRENCIA_ACC"] = killed_people_df["FECHA_HORA_ACC_r"].dt.year

    # Install the Spanish locale before running the following code
    # https://serverpilot.io/docs/how-to-install-locales/
killed_people_df["MES_OCURRENCIA_ACC"] = killed_people_df["FECHA_HORA_ACC_r"].dt.month_name(locale = "es_ES.UTF-8")
killed_people_df["MES_NRO_OCURRENCIA_ACC"] = killed_people_df["FECHA_HORA_ACC_r"].dt.month
killed_people_df["DIA_OCURRENCIA_ACC"] = killed_people_df["FECHA_HORA_ACC_r"].dt.day_name(locale = "es_ES.UTF-8")

    # 1 is Monday, 2 is Tuesday and so on
killed_people_df["DIA_NRO_OCURRENCIA_ACC"] = killed_people_df["FECHA_HORA_ACC_r"].dt.dayofweek + 1

killed_people_df["DIA_MES_OCURRENCIA_ACC"] = killed_people_df["FECHA_HORA_ACC_r"].dt.day
killed_people_df["HORA_OCURRENCIA_ACC"] = killed_people_df["FECHA_HORA_ACC_r"].dt.hour

# Uppercase month and day of week to keep consistency with original format
killed_people_df["MES_OCURRENCIA_ACC"] = killed_people_df["MES_OCURRENCIA_ACC"].str.upper()
killed_people_df["DIA_OCURRENCIA_ACC"] = killed_people_df["DIA_OCURRENCIA_ACC"].str.upper()

# Rename the longitude and latitude columns
killed_people_df.rename(columns = {"geometry.x": "LONGITUDE", "geometry.y": "LATITUDE"}, inplace = True)

# GENERO NaN treatment: fill with "SIN INFORMACION"
killed_people_df["GENERO"] = killed_people_df["GENERO"].fillna("SIN INFORMACION")

# MUERTE_POSTERIOR NaN treatment: fill with "N" (the person die in the accident, not after)
killed_people_df["MUERTE_POSTERIOR"] = killed_people_df["MUERTE_POSTERIOR"].fillna("N")

# EDAD NaN treatment: since we don't have information on what a zero represents (is it a baby that hasn't turned one year yet or
# a null value?), we leave this feature as it is

killed_people_df.to_json("shapefiles/killed_people_2015-2022_r2.json")

killed_people_df.info()

killed_people_df.head()

# Actors

actors_df = pd.read_json("shapefiles/actors/actors_2015-2022.json")

# GENERO NaN treatment: fill with "SIN INFORMACION"
actors_df["GENERO"] = actors_df["GENERO"].fillna("SIN INFORMACION")

# EDAD NaN treatment: since we don't have information on what a zero represents (is it a baby that hasn't turned one year yet or
# a null value?), we leave this feature as it is

# CODIGO_VEHICULO and CONDICION NaN treatment: PASAJERO, PEATON and null values in CONDICION are associated with null values in
# CODIGO_VEHICULO. We fill null values in CONDICION with "SIN INFORMACION" and null values in CODIGO_VEHICULO with 0s and then
# change their type to int
actors_df["CONDICION"] = actors_df["CONDICION"].fillna("SIN INFORMACION")
actors_df["CODIGO_VEHICULO"] = actors_df["CODIGO_VEHICULO"].fillna(0)
actors_df["CODIGO_VEHICULO"] = actors_df["CODIGO_VEHICULO"].astype(int) 

# ESTADO NaN treatment: fill with "ILESO"
actors_df["ESTADO"] = actors_df["ESTADO"].fillna("ILESO")

# Note:
# len(actors_df[(actors_df["ESTADO"] == "MUERTO")]) = 2738.
# len(actors_df[(actors_df["ESTADO"] == "HERIDO") & (actors_df["MUERTE_POSTERIOR"] == "S")]) = 1090.
# The sum should be 3829, which is the length of killed_people_df, but the sum is 3828.
# However, digging deeper into the actors dataset, there's one uninjured actor (ESTADO ILESO) with MUERTE_POSTERIOR S. Adding
# this actor, the sum is 3829

# MUERTE_POSTERIOR NaN treatment: this feature is related to ESTADO. There're some combinations that don't make much sense when
# compared to other combinations in the dataset (ESTADO MUERTO - MUERTE_POSTERIOR S, ESTADO HERIDO - MUERTE POSTERIOR S and the
# ILESO case mentioned above). It could be that ESTADO MUERTO - MUERTE_POSTERIOR S corresponds to ESTADO HERIDO - 
# MUERTE_POSTERIOR S.
# Given these inconsistencies, we don't further modify ESTADO and leave MUERTE_POSTERIOR as it is.
# However, we engineer a new feature called ESTADO_FINAL
actors_df["ESTADO_FINAL"] = np.where((actors_df["ESTADO"] == "MUERTO") | (actors_df["MUERTE_POSTERIOR"] == "S"), "MUERTO", \
    actors_df["ESTADO"])
# len(actors_df[(actors_df["ESTADO_FINAL"] == "MUERTO")]) = 3829.
# A thorough approach would have to check actors_df against accidents_df GRAVEDAD, injured_people_df and killed_people_df to fix
# inconsistencies if any

actors_df.to_json("shapefiles/actors_2015-2022_r1.json")

actors_df.info()

actors_df.head()

# Vehicles

vehicles_df = pd.read_json("shapefiles/vehicles/vehicles_2015-2022.json")

# CODIGO_VEHICULO NaN treatment: fill with 1s and then change its type to int
vehicles_df["CODIGO_VEHICULO"] = vehicles_df["CODIGO_VEHICULO"].fillna(1)
vehicles_df["CODIGO_VEHICULO"] = vehicles_df["CODIGO_VEHICULO"].astype(int)

# CLASE NaN treatment: fill with "SIN INFORMACION"
vehicles_df["CLASE"] = vehicles_df["CLASE"].fillna("SIN INFORMACION")

# SERVICIO NaN treatment: fill with "SIN INFORMACION"
# Do some cleaning in rows with values " SIN INFORMACION"
vehicles_df["SERVICIO"] = vehicles_df["SERVICIO"].fillna("SIN INFORMACION")
vehicles_df["SERVICIO"] = np.where(vehicles_df["SERVICIO"] == " SIN INFORMACION", "SIN INFORMACION", vehicles_df["SERVICIO"])

# MODALIDAD NaN treatment: fill with "SIN INFORMACION".
# While this could be done in a single step, we first checked with SERVICIO PUBLICO since the understanding was that MODALIDAD
# was a feature that only applied to SERVICIO PUBLICO vehicles
vehicles_df["MODALIDAD"] = np.where((vehicles_df["SERVICIO"] == "PUBLICO") & (vehicles_df["MODALIDAD"].isna()), \
    "SIN INFORMACION", vehicles_df["MODALIDAD"])
vehicles_df["MODALIDAD"] = np.where((vehicles_df["SERVICIO"] != "PUBLICO") & (vehicles_df["MODALIDAD"].isna()), \
    "SIN INFORMACION", vehicles_df["MODALIDAD"])

# ENFUGA NaN treatment: there're only 3 null values, but the info is rather confusing. So, we leave it as it is

vehicles_df.to_json("shapefiles/vehicles_2015-2022_r1.json")

vehicles_df.info()

vehicles_df.head()

# Causes

causes_df = pd.read_json("shapefiles/causes/causes_2015-2022.json")

# The NaN values are associated with an undetermined cause present in 11 rows. They are in TIPO and TIPO_CAUSA
causes_df.fillna("SIN ESTABLECER", inplace = True)

# Do some cleaning in rows with values "OTRAS"
causes_df["NOMBRE"] = np.where(causes_df["NOMBRE"] == "OTRAS", "OTRA", causes_df["NOMBRE"])

causes_df.to_json("shapefiles/causes_2015-2022_r1.json")

causes_df.info()

causes_df.head()
