==> initial_etl/4_database_creation.ipynb <==
import pandas as pd
import numpy as np
import psycopg2
import psycopg2.extras as extras

# We load all DataFrames

accidents_df = pd.read_json("shapefiles/accidents_2015-2022_r1.json")
injured_people_df = pd.read_json("shapefiles/injured_people_2015-2022_r1.json")
killed_people_df = pd.read_json("shapefiles/killed_people_2015-2022_r2.json")
vehicles_df = pd.read_json("shapefiles/vehicles_2015-2022_r1.json")
actors_df = pd.read_json("shapefiles/actors_2015-2022_r1.json")
causes_df = pd.read_json("shapefiles/causes_2015-2022_r1.json")

# Make sure to create the postgres user "dev" with password "dev" and the database "accidents_smb" to make this work.
# We create the tables in the database

# https://www.tutorialspoint.com/python_data_access/python_postgresql_database_connection.htm
db_conn = psycopg2.connect(
   database = "accidents_smb", user = "dev", password = "dev", host = "127.0.0.1", port = "5432"
)
cursor = db_conn.cursor()
commands = (
    """
    CREATE TABLE siniestros(
        OBJECTID INT,
        FORMULARIO VARCHAR(25) CONSTRAINT SINIESTROS_PK PRIMARY KEY,
        LOCALIDAD VARCHAR(50),
        CIV INT,
        PK_CALZADA INT,
        LONGITUD FLOAT(10),
        LATITUD FLOAT(10),
        CLASE_ACC VARCHAR(25),
        GRAVEDAD VARCHAR(25),
        FECHA_ACC VARCHAR(25),
        ANO_OCURRENCIA_ACC SMALLINT,
        MES_OCURRENCIA_ACC VARCHAR(25),
        MES_NRO_OCURRENCIA_ACC SMALLINT,
        DIA_OCURRENCIA_ACC VARCHAR(25),
        DIA_NRO_OCURRENCIA_ACC SMALLINT,
        DIA_MES_OCURRENCIA_ACC SMALLINT,
        HORA_OCURRENCIA_ACC SMALLINT
    )
    """,
    
    """
    CREATE TABLE conheridos(
        OBJECTID INT,
        FORMULARIO VARCHAR(25),
        LOCALIDAD VARCHAR(50),
        LONGITUD FLOAT(10),
        LATITUD FLOAT(10),
        CLASE_ACC VARCHAR(25),
        CONDICION VARCHAR(25),
        GENERO VARCHAR(25),
        EDAD FLOAT(1),
        FECHA_ACC VARCHAR(25),
        ANO_OCURRENCIA_ACC SMALLINT,
        MES_OCURRENCIA_ACC VARCHAR(25),
        MES_NRO_OCURRENCIA_ACC SMALLINT,
        DIA_OCURRENCIA_ACC VARCHAR(25),
        DIA_NRO_OCURRENCIA_ACC SMALLINT,
        DIA_MES_OCURRENCIA_ACC SMALLINT,
        HORA_OCURRENCIA_ACC SMALLINT,
        CONSTRAINT CONHERIDOS_FK
            FOREIGN KEY(FORMULARIO) 
                REFERENCES siniestros(FORMULARIO)
                ON DELETE CASCADE
    )
    """,
    
    """
    CREATE TABLE confallecidos(
        OBJECTID INT,
        FORMULARIO VARCHAR(25),
        LOCALIDAD VARCHAR(50),
        LONGITUD FLOAT(10),
        LATITUD FLOAT(10),
        CLASE_ACC VARCHAR(25),
        CONDICION VARCHAR(25),
        GENERO VARCHAR(25),
        EDAD FLOAT(1),
        MUERTE_POSTERIOR VARCHAR(5),
        FECHA_POSTERIOR_MUERTE FLOAT(1),
        FECHA_ACC VARCHAR(25),
        ANO_OCURRENCIA_ACC SMALLINT,
        MES_OCURRENCIA_ACC VARCHAR(25),
        MES_NRO_OCURRENCIA_ACC SMALLINT,
        DIA_OCURRENCIA_ACC VARCHAR(25),
        DIA_NRO_OCURRENCIA_ACC SMALLINT,
        DIA_MES_OCURRENCIA_ACC SMALLINT,
        HORA_OCURRENCIA_ACC SMALLINT,
        CONSTRAINT CONFALLECIDOS_FK
            FOREIGN KEY(FORMULARIO) 
                REFERENCES siniestros(FORMULARIO)
                ON DELETE CASCADE
    )
    """,
    
    """
    CREATE TABLE vehiculos(
        OBJECTID INT,
        FORMULARIO VARCHAR(25),
        CODIGO_VEHICULO SMALLINT,
        CLASE VARCHAR(25),
        SERVICIO VARCHAR(25),
        MODALIDAD VARCHAR(50),
        ENFUGA VARCHAR(5),
        CONSTRAINT VEHICULOS_FK
            FOREIGN KEY(FORMULARIO) 
                REFERENCES siniestros(FORMULARIO)
                ON DELETE CASCADE
    )
    """,
    
    """
    CREATE TABLE actores(
        OBJECTID INT,
        FORMULARIO VARCHAR(25),
        CODIGO_VICTIMA SMALLINT,
        CODIGO_VEHICULO	SMALLINT,
        CONDICION VARCHAR(25),
        GENERO VARCHAR(25),
        EDAD FLOAT(1),
        ESTADO VARCHAR(25),
        MUERTE_POSTERIOR VARCHAR(5),
        FECHA_POSTERIOR_MUERTE FLOAT(1),
        ESTADO_FINAL VARCHAR(25),
        CONSTRAINT ACTORES_FK
            FOREIGN KEY(FORMULARIO) 
                REFERENCES siniestros(FORMULARIO)
                ON DELETE CASCADE
    )
    """,
    
    """
    CREATE TABLE causas(
        OBJECTID INT,
        FORMULARIO VARCHAR(25),
        CODIGO_VEHICULO SMALLINT,
        CODIGO_CAUSA VARCHAR(10),
        NOMBRE VARCHAR(75),
        TIPO VARCHAR(25),
        TIPO_CAUSA VARCHAR(25),
        CONSTRAINT CAUSAS_FK
            FOREIGN KEY(FORMULARIO) 
                REFERENCES siniestros(FORMULARIO)
                ON DELETE CASCADE
    )
    """
)
for command in commands:
    cursor.execute(command)
cursor.close()
db_conn.commit()
db_conn.close()

# https://www.geeksforgeeks.org/how-to-insert-a-pandas-dataframe-to-an-existing-postgresql-table/

def execute_values(conn, df, table):
  
    tuples = [tuple(x) for x in df.to_numpy()]
  
    cols = ','.join(list(df.columns))
    # SQL query to execute
    query = "INSERT INTO %s(%s) VALUES %%s" % (table, cols)
    cursor = conn.cursor()
    try:
        extras.execute_values(cursor, query, tuples)
        conn.commit()
    except (Exception, psycopg2.DatabaseError) as error:
        print("Error: %s" % error)
        conn.rollback()
        cursor.close()
        return 1
    print("the dataframe is inserted")
    cursor.close()

# Make some adjustments on accidents_df

accidents_df["FECHA_ACC"] = pd.to_datetime(accidents_df["FECHA_HORA_ACC_r"], unit = "ms").dt.date
accidents_df.drop({"FECHA_HORA_ACC", "FECHA_HORA_ACC_r"}, axis = 1, inplace = True)
accidents_df.rename(columns = {"LONGITUDE": "LONGITUD", "LATITUDE": "LATITUD"}, inplace = True)
accidents_df.head()

# Insert the accidents DataFrame into the siniestros table

db_conn = psycopg2.connect(
   database = "accidents_smb", user = "dev", password = "dev", host = "127.0.0.1", port = "5432"
)
execute_values(db_conn, accidents_df, "siniestros")

# Make some adjustments on injured_people_df

injured_people_df["FECHA_ACC"] = pd.to_datetime(injured_people_df["FECHA_HORA_ACC_r"], unit = "ms").dt.date
injured_people_df.drop({"FECHA_HORA_ACC", "FECHA_HORA_ACC_r"}, axis = 1, inplace = True)
injured_people_df.rename(columns = {"LONGITUDE": "LONGITUD", "LATITUDE": "LATITUD"}, inplace = True)
injured_people_df.head()

# Insert the injured people DataFrame into the conheridos table

db_conn = psycopg2.connect(
   database = "accidents_smb", user = "dev", password = "dev", host = "127.0.0.1", port = "5432"
)
execute_values(db_conn, injured_people_df, "conheridos")

# Make some adjustments on killed_people_df

killed_people_df["FECHA_ACC"] = pd.to_datetime(killed_people_df["FECHA_HORA_ACC_r"], unit = "ms").dt.date
killed_people_df.drop({"FECHA_HORA_ACC", "FECHA_HORA_ACC_r"}, axis = 1, inplace = True)
killed_people_df.rename(columns = {"LONGITUDE": "LONGITUD", "LATITUDE": "LATITUD"}, inplace = True)
killed_people_df.head()

# Insert the killed people DataFrame into the confallecidos table

db_conn = psycopg2.connect(
   database = "accidents_smb", user = "dev", password = "dev", host = "127.0.0.1", port = "5432"
)
execute_values(db_conn, killed_people_df, "confallecidos")

# Insert the vehicles DataFrame into the vehiculos table

db_conn = psycopg2.connect(
   database = "accidents_smb", user = "dev", password = "dev", host = "127.0.0.1", port = "5432"
)
execute_values(db_conn, vehicles_df, "vehiculos")

# Insert the actors DataFrame into the actores table

db_conn = psycopg2.connect(
   database = "accidents_smb", user = "dev", password = "dev", host = "127.0.0.1", port = "5432"
)
execute_values(db_conn, actors_df, "actores")

# Insert the causes DataFrame into the causas table

db_conn = psycopg2.connect(
   database = "accidents_smb", user = "dev", password = "dev", host = "127.0.0.1", port = "5432"
)
execute_values(db_conn, causes_df, "causas")
